refactor: move schemaProperties setting logics to helper
